Add files via upload
Explains the imports of of three different classes. The tool class for utilization of various tools by the LLM/agents we introducing,  Additionally we break down the AgentType class for its reference to the OpenAI network for utilization of different agents for different tasks. Finally the OpenAIEmbeddings class which allows our agents to use context of previous data to adjust future data insights we are trying to extract.